fix layout structure of navbar
