+++ src/pages/Home.jsx
@@ -1,30 +1,15 @@
 import { useSelector } from "react-redux";
-import { useState, useEffect } from "react";
 
 import GameList from '../Componenets/GameList';
-import AppBar from '../Layout/Navbar';
 
-export default function Home() {
-
-  const done = useSelector((state) => state.data.done);
-  const list = useSelector((state) => state.data.items);
+export default function Home({games, Done}) {
   const favorites = useSelector((state) => state.data.favorites);
-  const [filteredArr, setfilter] = useState([]);
 
-  useEffect(() => {
-    setfilter([...list]);
-  }, [list]);
+  console.log(Done)
 
-  const searchHandler = (e) => {
-    const search = e.target.value;
-    const copy = list.filter((item) => item.title.includes(search));
-    setfilter(copy);
-  };
-  
   return (
     <>
-      <AppBar inputSearchHandler={searchHandler} />
-      <GameList Games={filteredArr} favorites={favorites} Done={done} />
+      <GameList Games={games} favorites={favorites} Done={Done} />
     </>
   )
 }
